Create DAY 14- REMOVE LINKED LIST ELEMENT (LINKED LIST)
QUESTION: Given the head of a linked list and an integer val, remove all the nodes of the linked list that has Node.val == val, and return the new head.

TESTCASES:
Case1: Input: head = [1,2,3,3,4,5,3], val = 3
           Output: [1,2,4,5]
Case 2: Input: head = [], val = 1
           Output: []
Case 3: Input: head = [0,0,0,0,] val = 0
            Output: []

EXAPLANATION:
1. We can use while loop for checking every element on the linked.
2. Inside while loop we will use if condition to check our the element present at that 
    position is equal to the value input. 

LOGIC: 
We can use a remove() Method 
It is used to remove an element from a linked list. The element is removed from the beginning or head of the linked list.
Syntax: 
LinkedList.remove()
Parameters: This function does not take any parameter.
Return Value: This method returns the head of the list or the element present at the head of the list.