Create DAY 15- MERGE TWO SORTED LIST (LINKED LIST)
Question: Merge the two lists in a one sorted list. The list should be made by splicing together the nodes of the first two lists. Return the head of the merged linked list. Return the head of the merged linked list.

TESYCASE1:
Input: list1 = [1,2,4,5], list2 = [1,3,4,6]
Output: [1,1,2,3,4,4,5,6]

TESTCASE2:
Input: list1 = [], list2 = []
Output: []

TESTCASE3:
Input: list1 = [], list2 = [0]
Output: [0]
 
EXAMPLE:
Your input: [1,4,6,9]
[5,7,9]
Output: [1,2,4,5,6,7,9,9]
Expected: [1,2,4,5,6,7,9,9]

Explanation:
The key to solve the problem is defining a fake head. Then compare the first elements from each elements list. Add the smaller one to the merged list. Finally, when one of them is empty, simply append it to the merged list, since it is already sorted.

Algorithms:
1) Create a new head pointer to an empty linked list.
(2) Check the first value of both linked lists.
(3) Whichever node from L1 or L2 is smaller, append it to the new list and move the pointer to the next node.
(4) Continue this process until you reach the end of a linked list.