Create DAY 16 - CONVERT BINARY NUMBER IN A LINKED LIST TO INTEGER(LINKED LIST)
Question: There is a reference node to a singly-linked list. The value of each node in the linked list is either 0 or 1. The linked list holds the binary representation of a number. Return the decimal value of the number in the linked list. The most significant bit is at the head of the linked list. 

TESTCASE1:
Input: head = [1,0,1]
Output: 5
Explanation: (101) in base 2 = (5) in base 10

TESTCASE2:
Input: head = [0]
Output: 0
 
EXAMPLE:
Your input: [1,1,1]
Output: 7
Expected: 7

Given Constraints:
The Linked List is not empty.
Number of nodes will not exceed 30.
Each node's value is either 0 or 1.

Explanation:
Here we have two sub-problems:
To parse non-empty linked list and to retrieve the digit sequence which represents a binary number. 
To convert this sequence into the number in decimal representation.