Create DAY 17 - MIDDLE OF THE LINKED LIST (LINKED LIST)
Given the head of a singly linked list, return the middle node of the linked list.
If there are two middle nodes, return the second middle node.

TESTCASE 1: Input: head = [9,10,11,12,13]
Output: [11,12,13]
Explanation: The middle node of the list is node 11.


TESTCASE2:  Input: head = [3,4,5,6,7,8]
Output: [5,6]
Explanation: Since the list has two middle nodes with values 5 and 6, we return the second one.

EXAMPLE:
 If the given linked list is 1->2->3->4->5 then the output should be 3. 
If there are even nodes, then there would be two middle nodes, we need to print the second middle element. For example, if the given linked list is 11>12->13->14->15->16 then the output should be14. 

LOGIC:
1. Traverse linked list using two-pointers. 
2. Move one pointer by one and the other pointers by two.
3. When the fast pointer reaches the end, the slow pointer will reach the middle of the linked list.