Create DAY 20- PALINDROME LINKED LIST (LINKED LIST)
Given the head of a singly linked list, return true if it is a palindrome or false otherwise.

TESTCASE:
CASE1:
Input: head = [1,2,3,32,1]
Expected: true
Output: true

CASE2:
Input: head = [1,2,5]
Output: false
Expected: false
LOGIC:
Determine if the list is a palindrome. We can create a new list in reversed order and then compare each node. The time and space are O(n). We can use a fast and slow pointer to get the center of the list, then reverse the second list and compare two sublists.