Create DAY 21 -  ROTATE LINKED (LINKED LIST)
Given the head of a linked list, rotate the list to the right by k places.

TESTCASES:
CASE1:
Input: head = [1,2,3,4,5], k = 2
Output: [4,5,1,2,3]

CASE2:
Input: head = [0,1,2], k = 4
Output: [2,0,1]

LOGIC:
We use the rotate() method of the Collections class to perform the right Rotation in the list. The method accepts two parameters, i.e., a list and a positive integer. List parameter defines a list in which we need to perform Rotation. The distance parameter is a positive integer that defines the number of Rotations.