Create DAY 40 - LONGEST HARMONIOUS SUBSEQUENCE (MAP)
Problem: We define a harmonious array as an array where the difference between its maximum value and its minimum value is exactly 1. Given an integer array nums, return the length of its longest harmonious subsequence among all its possible subsequences. A subsequence of array is a sequence that can be derived from the array by deleting some or no elements without changing the order of the remaining elements.
 
Example 1:
Input: nums = [1,3,2,2,5,2,3,7]
Output: 5
Explanation: The longest harmonious subsequence is [3,2,2,2,3].

Example 2:
Input: nums = [1,2,3,4]
Output: 2

Testcase:
Your input
[1,3,2,2,5,2,3,7]
Output
5
Expected
5

Explanaton:
In this approach, we make use of a hashmap map which stores the number of times an element occurs in the array along with the element's value in the form (num, count_num), where num refers to an element in the array and count_num refers to the number of times this num occurs in the nums nums array. We traverse over the nums array and fill this map once. After this, we traverse over the keys of the map created.  For every key of the map considered, say key, we find out if the map contains the key+1.