Create DAY 58 - MULTIPLY STRINGS (STRINGS)
PROBLEM: Given two non-negative integers num1 and num2 represented as strings, return the product of num1 and num2, also represented as a string.

EXAMPLE
Example 1:
Input: num1 = "2", num2 = "3"
Output: "6"

TESTCASES
CASE1:
Input: num1 ="123" num2 ="456"
Output: "56088"
Expected: "56088"

CASE2:
Input: num1 ="3" num2 ="0"
Output: "0"
Expected: "0"

CASE3:
Input: num1 ="3" num2 ="4"
Output: "12"
Expected: "12"

EXPLANATION:
So, in this approach, we will traverse over one of the given strings from right to left. When we are at an index or a character, we multiply whole of the other string with this current character. Once we are done with our multiplication, we add zeroes to the end of it. The number of zeroes that are to be added equals the position of the current character in its string from the end – 1. Once we are done with adding zeroes, we add this string to the result. So, as we have traversed from right to left of the first string, the resultant string stores the answer.