Create DAY 60 - COUSINS IN BINARY TREE (BINARY TREE)
PROBLEM: In a binary tree, the root node is at depth 0, and children of each depth k node are at depth k+1.
Two nodes of a binary tree are cousins if they have the same depth, but have different parents.

EXAMPLES
Example 1:
Input: root = [1,2,3,4], x = 4, y = 3
Output: false

TESTCASES:
CASE 1:
Input: root = [1,2,3,null,4,null,5], x = 5, y = 4
Output: true

CASE 2:
Input: root = [1,2,3,null,4], x = 2, y = 3
Output: false

EXPLANATION:
Since cousins are defined to be nodes at the same level of a tree, we resort to queue as our main data structure. Insert NULL to the queue when we reach the very end of each level. Also we need a way to make sure the nodes we find that have the expected values do not have the same parent. How to achieve this without maintaining extra parent information per node? Here is what we do: if the left child has one of the expected values, we do not check the right child and vice versa. In this way if both children have the expected values we only count one occurrence instead of two occurrences.