Create DAY 67 - LONGEST SUBSTRING WITHOUT REPEATING CHARACTERS (STRING)
PROBLEM: Given a string s, find the length of the longest substring without repeating characters.

EXAMPLE
Example 1:
Example 1:
Input: s = "abcabcbb"
Output: 3
Explanation: The answer is "abc", with the length of 3.

TESTCASES
Case1: 
Input s ="bbbbb"
Output: 1
Expected: 1

Case2:
Input s ="pwwkew"
Output: 3
Expected: 3

EXPLANATION:
Following are the steps that we will follow:
A. Have two pointers which will define the starting index start and ending index end of the current window. Both will be 0 at the beginning.
B. Declare a Set that will store all the unique characters that we have encountered.
C. Declare a variable maxLength that will keep track of the length of the longest valid substring.
D. Scan the string from left to right one character at a time.
E. If the character has not encountered before i.e., not present in the Set the we will add it and increment the end index. The maxLength will be the maximum of Set.size() and existing maxLength.
F. If the character has encounter before, i.e., present in the Set, we will increment the start and we will remove the character at start index of the string.
G. After the loop terminates, return maxLength.