Create DAY 77 LINKED LIST CYCLE (LINKED LIST)
PROBLEM: Given head, the head of a linked list, determine if the linked list has a cycle in it.

EXAMPLE
Example1:
Input: head = [3,2,0,-4], pos = 1
Output: true
Explanation: There is a cycle in the linked list, where the tail connects to the 1st node (0-indexed).

TESTCASES
Case1:
Input: [1,2] 0
Output: true
Expected: true

Case2:
Input: [1] -1
Output: false
Expected: false

EXPLANATION:
The brute-force solution uses a HashSet. Assuming how a basic HashSet works according to the Collections framework. We traverse through the entire list starting from the head of the list and add every new node encountered in to the list . Incase we come into a node that has already been encountered in the list then we return false. On reaching the last node if all nodes of the list are unique true is returned.