Create DAY 80 - POWER OF TWO
PROBLEM: Given an integer n, return true if it is a power of two. Otherwise, return false.
An integer n is a power of two, if there exists an integer x such that n == 2x.

EXAMPLE
Example1:
Input: n = 1
Output: true
Explanation: 20 = 1

TESTCASES
Case1:
Input: n = 16
Output: true
Explanation: 24 = 16
Example 3:

Case2:
Input: n = 3
Output: false

EXPLANATION:
If a number is power of 2, it's binary form should be 10...0. So if we right shift a bit of the number and then left shift a bit, the value should be the same when the number >= 10(i.e.,2).