Create DAY 82 - REVERSE BITS
PROBLEM: Reverse bits of a given 32 bits unsigned integer.

EXAMPLE:
Example 1:

Input: n = 00000010100101000001111010011100
Output:    964176192 (00111001011110000010100101000000)
Explanation: The input binary string 00000010100101000001111010011100 represents the unsigned integer 43261596, so return 964176192 which its binary representation is 00111001011110000010100101000000.

TESTCASES 
CASE1:
Input: 11111111111111111111111111111101
Output: 3221225471 (10111111111111111111111111111111)
Expected:  3221225471 (10111111111111111111111111111111)

EXPLANATION:
Iterate through the number from LSB to MSB and flip the bits till there. But here we are dealing with the unsigned numbers and things get a bit tricky with unsigned numbers, and the numbers are given as the binary representation of the number too.