Create DAY 87 - DETECT CAPITAL (STRING)
PROBLEM: We define the usage of capitals in a word to be right when one of the following cases holds.

EXAMPLE
Example1:
Input: word = "USA"
Output: true

TESTCASE
Case1:
Input word ="FlaG"
Output: false
Expected: false

EXPLANATION:
First, take two variables to say, “capital” and “small” to store the count of the capital letters and the small letters which are present in the given string. Now iterate through the string and update the values accordingly, which means if a capital letter is found then increase the count of the “capital” variable otherwise increase the “small” variable.
Then check the first two conditions. If the length of the given string is equal to any of the variable “capital” or “small”, return true as all the letters of the word may be capital or may be small.
After that, check for the third given condition where first check if the first character of the string is a capital letter or not and the length of the remaining string is equal to the “small” variable or not. If both of them are true, return true.