Create DAY 88 - SUBTREE OF ANOTHER TREE (BINARY TREE)
PROBLEM: Given the roots of two binary trees root and subRoot, return true if there is a subtree of root with the same structure and node values of subRoot and false otherwise.

EXAMPLE
Example1:
Input: root = [3,4,5,1,2], subRoot = [4,1,2]
Output: true

TESTCASES
Case1:
Input root =[3,4,5,1,2,null,null,null,null,0] subRoot =[4,1,2]
Output: false
Expected: false

EXPLANATION:
When checking isSame, DFS in left and right seperately, and include both of them to draw the conclusion. For the isSubTree, when root heads to candidate node, to return true, all the rest decendants need to be identical with the tree of subroot. In other words, just use isSame to judge.