Create DAY 95 - LENGTH OF LAST WORD
PROBLEM: Given a string s consisting of words and spaces, return the length of the last word in the string.
A word is a maximal substring consisting of non-space characters only.

EXAMPLE
Example1:
Input: s = "Hello World"
Output: 5
Explanation: The last word is "World" with length 5.

TESTCASES
Case1:
Input s ="   fly me   to   the moon  "
Output: 4
Expected: 4

Case2:
Input s ="luffy is still joyboy"
Output: 6
Expected: 6

EXPLANATION:
To solve this problem first we have to trim all the whitespaces present in the string at both the ends. Then we can find the length of the last word by iterating the string from first character to last character. For trimming we can use two variables ‘start’ and ‘end’ which will store the index of first character and last character in the given string respectively.