Create DAY 97 - REMOVE OUTERMOST PARENTHESES
PROBLEM:  A valid parentheses string is either empty "", "(" + A + ")", or A + B, where A and B are valid parentheses strings, and + represents string concatenation.

EXAMPLE
Example1:
Input: s = "(()())(())"
Output: "()()()"
Explanation: 
The input string is "(()())(())", with primitive decomposition "(()())" + "(())".
After removing outer parentheses of each part, this is "()()" + "()" = "()()()".

TESTCASE
Case1:
Input: s = "(()())(())(()(()))"
Output: "()()()()(())"
Explanation: 
The input string is "(()())(())(()(()))", with primitive decomposition "(()())" + "(())" + "(()(()))".
After removing outer parentheses of each part, this is "()()" + "()" + "()(())" = "()()()()(())".

Case2:
Input: s = "()()"
Output: ""
Explanation: 
The input string is "()()", with primitive decomposition "()" + "()".
After removing outer parentheses of each part, this is "" + "" = "".

EXPLANATION:
Intuition is that we will iterate through each character and when we encounter opening parentheses we push to stack. while pushing to the stack we will check if there are at least 1 open parentheses and if that is the case that means we already covered outer parentheses so current open parentheses are part of the output.
If the encounter string is closing parentheses then we again check if the stack size is greater than 1. please notice here we need at least more than 1 opening parentheses in the stack so that we can consider adding closing parentheses to result.