Create DAY 99 - CONSTRUCT STRING FROM BINARY TREE
PROBLEM: Given the root of a binary tree, construct a string consisting of parenthesis and integers from a binary tree with the preorder traversal way, and return it.

EXAMPLE
Example1:
Input: root = [1,2,3,4]
Output: "1(2(4))(3)"
Explanation: Originally, it needs to be "1(2(4)())(3()())", but you need to omit all the unnecessary empty parenthesis pairs. And it will be "1(2(4))(3)"

TESTCASES
Case1:
Input root =[1,2,3,null,4]
Output: "1(2()(4))(3)"
Expected: "1(2()(4))(3)"

EXPLANATION:
Traverse the tree and append the values and parenthesis. If the node doesn’t have left and right subtrees, we don’t need to append parenthesis. If the node only has right subtrees, we should append parenthesis as a left subtree.