Create DAY 100 - RECOVER BINARY SEARCH TREE
PROBLEM: You are given the root of a binary search tree (BST), where the values of exactly two nodes of the tree were swapped by mistake. Recover the tree without changing its structure.

EXAMPLE
Example1:
Input: root = [1,3,null,null,2]
Output: [3,1,null,null,2]
Explanation: 3 cannot be a left child of 1 because 3 > 1. Swapping 1 and 3 makes the BST valid.

TESTCASES
Case1:
Input root =[3,1,4,null,null,2]
Output: [2,1,4,null,null,3]
Expected: [2,1,4,null,null,3]

EXPLANATION:
The main idea to solve this problem is to use Inorder Traversal of the Binary Search Tree.
In an Inorder traversal of a Binary Search Tree, It generated the sequence of node’s values in increasing order.
Now, Consider when all the node values are at their correct place in a Binary Search Tree. Let’s find the prefix maximum of such sorted sequence. Then, we can easily observe that for every integer in the sequence prefix maximum [i] = node value at the current position.
When two node values are swapped, then there exists a subarray where the above criteria don’t hold. Hence, we need to find the starting and ending positions of such subarray.
Swap the node values at this two-position and our binary search tree is recovered.
Perform the inorder traversal of the tree in a recursive fashion and find the starting and ending node of the required subarray.
Swap the two-node values found in the above step.